# Conflicts:
#	src/main/java/ku/cs/usecasedesigner/controller/HomePageController.java

+++ src/main/java/ku/cs/usecasedesigner/models/Position.java
@@ -2,14 +2,14 @@ package ku.cs.usecasedesigner.models;
 
 public class Position {
     private String type;
-    private double position_id;
-    private double symbol_id;
+    private int position_id;
+    private int symbol_id;
     private double x_position;
     private double y_position;
     private double fit_width;
     private double fit_height;
     private double rotation;
-    public Position(double position_id, double symbol_id, double x_position, double y_position, double fit_width, double fit_height, double rotation) {
+    public Position(int position_id, int symbol_id, double x_position, double y_position, double fit_width, double fit_height, double rotation) {
         this.type = "position";
         this.position_id = position_id;
         this.symbol_id = symbol_id;
@@ -20,11 +20,11 @@ public class Position {
         this.rotation = rotation;
     }
 
-    public double getPosition_id() {
+    public int getPosition_id() {
         return position_id;
     }
 
-    public double getSymbol_id() {
+    public int getSymbol_id() {
         return symbol_id;
     }
