+++ src/main/java/ku/cs/usecasedesigner/models/SymbolList.java
@@ -39,8 +39,8 @@ public class SymbolList {
         return null;
     }
 
-    public double findLastSymbolId() {
-        double lastSymbolId = 0;
+    public int findLastSymbolId() {
+        int lastSymbolId = 0;
         for (Symbol symbol : symbolList) {
             if (symbol.getSymbol_id() > lastSymbolId) {
                 lastSymbolId = symbol.getSymbol_id();
